+++ src/index.js
@@ -6,16 +6,8 @@ import './javascript/responseParser';
 import './javascript/mainPageWeatherInfo';
 import './javascript/main';
 
-import phrases from './phrases';
-import randomNumberSelection from './randomNumberSelection.js';
-import markup from './template/phrasesTemplate.hbs';
-import dateTimeLoader from "./date-time-info/loadDate";
 
-const markupForPhrases = function (markup) {
-  const randomPhrases = phrases[randomNumberSelection(0, phrases.length - 1)];
-  const markupPhrases = markup(randomPhrases);
-  document
-    .querySelector('.quote-wrap')
-    .insertAdjacentHTML('beforeend', markupPhrases);
-};
-markupForPhrases(markup);
+import './phrases/phrases-markup';
+import "./date-time-info/loadDate";
+
+
